requirements.txt
astroid==2.3.3
attrs==19.3.0
awsebcli==3.17.1
bcrypt==3.1.7
blessed==1.17.2
botocore==1.14.17
cached-property==1.5.1
cement==2.8.2
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
colorama==0.3.9
cryptography==2.8
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.10
django-daterange-filter==1.3.0
django-import-export==2.0.1
docker==4.2.0
docker-compose==1.25.4
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
et-xmlfile==1.0.1
future==0.16.0
idna==2.7
importlib-metadata==1.5.0
isort==4.3.21
jdcal==1.4.1
jmespath==0.9.4
jsonschema==3.2.0
lazy-object-proxy==1.4.3
MarkupPy==1.14
mccabe==0.6.1
odfpy==1.4.1
openpyxl==2.6.4
paramiko==2.7.1
pathspec==0.5.9
pycparser==2.19
pylint==2.4.4
PyNaCl==1.3.0
pyrsistent==0.15.7
python-dateutil==2.8.0
pytz==2019.3
PyYAML==5.2
requests==2.20.1
semantic-version==2.5.0
six==1.11.0
sqlparse==0.3.0
tablib==0.14.0
termcolor==1.1.0
texttable==1.6.2
typed-ast==1.4.1
urllib3==1.24.3
wcwidth==0.1.8
websocket-client==0.57.0
wrapt==1.11.2
xlrd==1.2.0
xlwt==1.3.0
zipp==1.2.0
